Added option to load hr file from wannier90 format.
Added option to save hr data in wannier90 format.

Chaning the config flag tp_params to hr_input which is a list if type = t_tp_tpp or a string for the input file it type='from_wannier90'

Ek generation is now done by the LatticeConfig.

Introduced the WannierHr class which handles ek generation.

hr and hk have been merged to the wannier.py module. The main code runs locally without complains, but it is expected that several other files (testing, postproc, etc.) still need an update to the new structure.